+++ protected/humhub/modules/activity/migrations/m150703_130157_migrate.php
@@ -15,7 +15,7 @@ class m150703_130157_migrate extends Migration
         // Space Created Activity - object_model/object_id (source fix)
         $activities = (new Query())->select('activity.*, content.space_id')->from('activity')
             ->leftJoin('content', 'content.object_model=:activityModel AND content.object_id=activity.id', [':activityModel' => Activity::class])
-            ->where(['class' => 'humhub\modules\space\activities\Created', 'activity.object_model' => ''])->all();
+            ->where(['class' => 'humhub\modules\space\activities\SpaceCreatedActivity', 'activity.object_model' => ''])->all();
         foreach ($activities as $activity) {
             $this->updateSilent('activity', [
                 'object_model' => Space::class,
@@ -26,7 +26,7 @@ class m150703_130157_migrate extends Migration
         // Space Member added Activity - object_model/object_id (source fix)
         $activities = (new Query())->select('activity.*, content.space_id')->from('activity')
             ->leftJoin('content', 'content.object_model=:activityModel AND content.object_id=activity.id', [':activityModel' => Activity::class])
-            ->where(['class' => 'humhub\modules\space\activities\MemberAdded', 'activity.object_model' => ''])->all();
+            ->where(['class' => 'humhub\modules\space\activities\MemberAddedActivity', 'activity.object_model' => ''])->all();
         foreach ($activities as $activity) {
             $this->updateSilent('activity', [
                 'object_model' => Space::class,
@@ -37,7 +37,7 @@ class m150703_130157_migrate extends Migration
         // Space Member removed Activity - object_model/object_id (source fix)
         $activities = (new Query())->select('activity.*, content.space_id')->from('activity')
             ->leftJoin('content', 'content.object_model=:activityModel AND content.object_id=activity.id', [':activityModel' => Activity::class])
-            ->where(['class' => 'humhub\modules\space\activities\MemberRemoved', 'activity.object_model' => ''])->all();
+            ->where(['class' => 'humhub\modules\space\activities\MemberRemovedActivity', 'activity.object_model' => ''])->all();
         foreach ($activities as $activity) {
             $this->updateSilent('activity', [
                 'object_model' => Space::class,