+++ protected/humhub/modules/like/services/LikeService.php
@@ -3,10 +3,11 @@
 namespace humhub\modules\like\services;
 
 use humhub\models\RecordMap;
+use humhub\modules\activity\services\ActivityManager;
 use humhub\modules\content\components\ContentAddonActiveRecord;
 use humhub\modules\content\interfaces\ContentProvider;
 use humhub\modules\content\models\Content;
-use humhub\modules\like\activities\Liked as LikedActivity;
+use humhub\modules\like\activities\LikeActivity as LikedActivity;
 use humhub\modules\like\models\Like;
 use humhub\modules\like\notifications\NewLike as NewLikeNotification;
 use humhub\modules\like\permissions\CanLike;
@@ -94,9 +95,9 @@ class LikeService
                 $this->reset();
 
                 $author = $this->contentAddon->createdBy ?? $this->content->createdBy;
-
                 NewLikeNotification::instance()->from($this->user)->about($record)->send($author);
-                LikedActivity::instance()->about($record)->save();
+
+                ActivityManager::dispatch(LikedActivity::class, $record, $record->createdBy);
 
                 return true;
             }